requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
flask-wtf==1.1.1
bootstrap-flask==2.2.0
